Reformat bash code

+++ reset-ccbd.sh
@@ -13,27 +13,35 @@ DOCKER_IMG="spark-hadoop"
 DOCKER_TAG="latest"
 
 usage() {
-	echo "Usage: sudo ./reset-ccbd.sh [OPTION]"
-	echo -e "\nReset the CCBD containers."
-	echo -e "\nOptions:"
-	echo -e "\t-h\tThe help menu"
-	echo -e "\t-i\tRemove the docker image"
+    echo "Usage: sudo ./reset-ccbd.sh [OPTION]"
+    echo -e "\nReset the CCBD containers."
+    echo -e "\nOptions:"
+    echo -e "\t-h\tThe help menu"
+    echo -e "\t-i\tRemove the docker image"
 }
 
-echo "stop all containers..."
-sudo docker stop $(docker ps -a -q) &> /dev/null
+init() {
+    echo "stop all containers..."
+    sudo docker stop $(docker ps -a -q) &> /dev/null
 
-while getopts ":hi" opt; do
-	case ${opt} in
-		h ) usage
-		  ;;
-		i ) echo "remove docker image..."
-                    sudo docker rmi -f $DOCKER_IMG:$DOCKER_TAG 2> /dev/null
-		  ;;
-		\? ) echo "Invalid option: $OPTARG" 1>&2
-		  ;;
-	esac
-done
+    echo "remove all docker network..."
+    sudo docker network rm $(docker network ls -q) &> /dev/null
+}
 
-echo "remove all docker network..."
-sudo docker network rm $(docker network ls -q) &> /dev/null
+if [[ ! $@ =~ ^\-.+ ]]; then
+    init
+else
+    while getopts ":hi" opt; do
+        case ${opt} in
+            h ) usage
+              ;;
+            i ) init
+                echo "remove docker image..."
+                sudo docker rmi -f $DOCKER_IMG:$DOCKER_TAG 2> /dev/null
+              ;;
+            \? ) echo -e "Invalid option: $OPTARG\n" 1>&2
+                 usage
+              ;;
+        esac
+    done
+fi
